# Improve performance of resolving manifest file paths