### Short description 📝

Running `tuist edit` in a large git repository is quite slow at the moment. Running Time Profiler indicates that globing is the root cause of the slowness as it globs the entire repository. This is despite `.tuistignore` being present, as exclusion is done only after globing.

These changes reduce the time it takes to run `tuist edit` in our large repository from 31s to ~3s~ ~1.5s~ 0.27s in release mode (M2 Max). Related discussion on [Tuist Slack](https://tuistapp.slack.com/archives/CCYNEGY1L/p1706843771208039).

I tried running `mise run tuist:test` but it fails acceptance tests even on a clean checkout. Could you advise on how tests should be run locally? I don't know if it's because of Tuist 4 changes, but a lot of the documentation seems a bit outdated on this.

**Before:**

<img width="655" alt="image" src="https://github.com/tuist/tuist/assets/133648611/43397b40-748c-4398-977b-45ada6dcfa2e">

**After:**

<img width="797" alt="image" src="https://github.com/tuist/tuist/assets/133648611/2fc45cc2-da83-457d-8390-078ac3e04910">

### How to test the changes locally 🧐

Run `tuist` scheme with `edit` argument against a custom working directory that is a large git repository.

### Contributor checklist ✅

- [x] The code has been linted using run `mise run lint:fix`
- [x] The change is tested via unit testing or acceptance testing, or both
- [x] The title of the PR is formulated in a way that is usable as a changelog entry
- [x] In case the PR introduces changes that affect users, the documentation has been updated

### Reviewer checklist ✅

- [ ] The code architecture and patterns are consistent with the rest of the codebase
- [ ] Reviewer has checked that, if needed, the PR includes the label `changelog:added`, `changelog:fixed`, or `changelog:changed`, and the title is usable as a changelog entry